Python Script/requirements.txt
matplotlib
numpy
pandas
opencv-python
scikit-learn
scikit-image
keras
tensorflow
torch
Pillow
scipy
statistics
tqdm
IPython
ipykernel